add controller for categoria, sexo, idioma and editorial

+++ src/main/java/com/rikazzo/back/repository/SexoRepository.java
@@ -3,6 +3,6 @@ package com.rikazzo.back.repository;
 import com.rikazzo.back.entity.Sexo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SexoRepository extends JpaRepository<Sexo, Integer> {
+public interface SexoRepository extends JpaRepository<Sexo, Short> {
 
 }
